requirements.txt
numpy==2.3.0
openpyxl==3.1.5
pandas==2.3.0
python-dotenv==1.1.0
streamlit==1.45.1
plotly

